backend/requirements.txt
Flask
Flask-CORS
python-dotenv
supabase
flask_jwt_extended
requests
